Add image to the user view
