+++ Firebase/android/app/src/main/java/com/novoda/bonfire/chat/view/MessageView.java
@@ -1,12 +1,6 @@
 package com.novoda.bonfire.chat.view;
 
 import android.content.Context;
-import android.graphics.Bitmap;
-import android.graphics.BitmapShader;
-import android.graphics.Canvas;
-import android.graphics.Paint;
-import android.graphics.RectF;
-import android.graphics.Shader;
 import android.view.Gravity;
 import android.view.View;
 import android.widget.ImageView;
@@ -14,10 +8,9 @@ import android.widget.LinearLayout;
 import android.widget.TextView;
 
 import com.bumptech.glide.Glide;
-import com.bumptech.glide.load.engine.bitmap_recycle.BitmapPool;
-import com.bumptech.glide.load.resource.bitmap.BitmapTransformation;
 import com.novoda.bonfire.R;
 import com.novoda.bonfire.chat.data.model.Message;
+import com.novoda.bonfire.view.CircleCropImageTransformation;
 import com.novoda.notils.caster.Views;
 
 import java.text.DateFormat;
@@ -32,6 +25,7 @@ public class MessageView extends LinearLayout {
     private final TextView body;
     private final TextView time;
     private final TextView name;
+    private CircleCropImageTransformation circleCropTransformation;
 
     public MessageView(Context context) {
         super(context);
@@ -41,6 +35,7 @@ public class MessageView extends LinearLayout {
         this.body = Views.findById(this, R.id.messageBody);
         this.time = Views.findById(this, R.id.messageTime);
         this.name = Views.findById(this, R.id.messageAuthorName);
+        circleCropTransformation = new CircleCropImageTransformation(context);
     }
 
     private void configureViewParams() {
@@ -55,7 +50,7 @@ public class MessageView extends LinearLayout {
     public void display(Message message) {
         Context context = getContext();
         Glide.with(context).load(message.getAuthor().getPhotoUrl())
-                .transform(new CircleCropTransformation(context))
+                .transform(circleCropTransformation)
                 .into(picture);
         body.setText(message.getBody());
         time.setText(formattedTimeFrom(message.getTimestamp()));
@@ -67,40 +62,4 @@ public class MessageView extends LinearLayout {
         return timeFormat.format(date);
     }
 
-    private static class CircleCropTransformation extends BitmapTransformation {
-
-        private final Paint paint = new Paint();
-
-        public CircleCropTransformation(Context context) {
-            super(context);
-        }
-
-        @Override
-        protected Bitmap transform(BitmapPool pool, Bitmap toTransform, int outWidth, int outHeight) {
-            Bitmap bitmap = getBitmap(pool, outWidth, outHeight);
-
-            BitmapShader shader = new BitmapShader(toTransform, Shader.TileMode.CLAMP, Shader.TileMode.CLAMP);
-
-            paint.setAntiAlias(true);
-            paint.setShader(shader);
-
-            RectF rect = new RectF(0.0f, 0.0f, outWidth, outHeight);
-            Canvas canvas = new Canvas(bitmap);
-            canvas.drawRoundRect(rect, outWidth / 2.0f, outHeight / 2.0f, paint);
-            return bitmap;
-        }
-
-        private Bitmap getBitmap(BitmapPool pool, int outWidth, int outHeight) {
-            Bitmap bitmap = pool.get(outWidth, outHeight, Bitmap.Config.ARGB_8888);
-            if (bitmap == null) {
-                bitmap = Bitmap.createBitmap(outWidth, outHeight, Bitmap.Config.ARGB_8888);
-            }
-            return bitmap;
-        }
-
-        @Override
-        public String getId() {
-            return "CircleCropTransformation";
-        }
-    }
 }
