+++ Firebase/android/app/src/main/res/layout/merge_user_item_view.xml
@@ -4,6 +4,13 @@
   android:layout_height="match_parent"
   android:orientation="horizontal">
 
+  <ImageView
+    android:id="@+id/userImage"
+    android:layout_width="40dp"
+    android:layout_height="40dp"
+    android:layout_marginEnd="16dp"
+    android:layout_marginRight="16dp" />
+
   <TextView
     android:id="@+id/userName"
     android:layout_width="wrap_content"
